# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Discussion Board API",
    "description": "Documentation API. https://github.com/danielschmitz/discussionboard-api <br/><br/>",
    "version": "1.0",
    "contact": {
      "name": "Daniel Schmitz",
      "email": "danieljfa@gmail.com",
      "url": "https://github.com/danielschmitz"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header"
    }
  },
  "paths": {
    "/hello-world": {
      "get": {
        "tags": [
          "Hello World"
        ],
        "summary": "A simple Hello World",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hello-world-auth": {
      "get": {
        "tags": [
          "Hello World"
        ],
        "summary": "🔒️ A simple Hello World with auth",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/info": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get info about token",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/isAdmin": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Check user is Admin",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Try to login",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User Login Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user1@email.com"
                },
                "password": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token"
          },
          "401": {
            "description": "Incorrect password"
          },
          "404": {
            "description": "No user found with that email"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get All Users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserResult"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "🔒️ Edit a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a user with that email"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "🔒️ Delete a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new User",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a email with that name"
          }
        }
      }
    },
    "/user/changePassword/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "🔒️ Change a password",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password updated successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "UserResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "User Name"
        },
        "email": {
          "type": "string",
          "example": "user@mail.com"
        },
        "isAdmin": {
          "type": "number",
          "example": 0
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "User Name"
        },
        "email": {
          "type": "string",
          "example": "user@mail.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}